requirements.txt
africastalking==1.2.5
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
contextlib2==21.6.0
Django==3.2.19
django-jazzmin==2.6.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.4.1
Markdown==3.4.3
pycodestyle==2.10.0
python-dotenv==0.21.1
pytz==2023.3
requests==2.29.0
schema==0.7.5
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
zipp==3.15.0
ptvsd
